Move `Drug` from `drug` to `drug.search` package
