+++ app/src/main/java/org/simple/clinic/drugs/search/Drug.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.drugs
+package org.simple.clinic.drugs.search
 
 import androidx.room.Embedded
 import androidx.room.Entity
